geotiff tests: consolidate writer-tail kwarg/shape validation (#2433)
Sub-PR A of cluster 9 (long-tail epic #2424). Fold five top-level
writer-flavoured validation files into write/test_basic.py:

- test_lowlevel_write_pushdown_2138.py
- test_to_geotiff_3d_dim_validation_1812.py
- test_temporal_3d_writer_rejection_1972.py
- test_to_geotiff_empty_shape_2075.py
- test_to_geotiff_zero_bands_2095.py

Tests-only. Module-level helpers that would collide with the host file
are suffixed per issue (e.g. _make_da_1812). The per-file GPU probes
collapse onto the host file's _gpu_only / @requires_gpu marker.

CLUSTER_AUDIT_WRITER_TAIL.md maps every old file::test to its new
file::test_id; it is deleted on a pre-merge commit.
